requirements.txt
distro==1.4.0
numpy==1.18.1
pandas==1.0.1
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
tabula-py==2.0.4
